[cpp] Backport of some 4.2 changes to enable Flutter runtime in 4.1

+++ spine-cpp/spine-cpp/include/spine/Vertices.h
@@ -35,7 +35,7 @@
 namespace spine {
 	class SP_API Vertices : public SpineObject {
 	public:
-		Vector <size_t> _bones;
+		Vector <int> _bones;
 		Vector<float> _vertices;
 	};
 }
